+++ tests/test_facade/test_psmac_facade.py
@@ -1,3 +1,4 @@
+import os
 import glob
 import shutil
 import unittest
@@ -22,30 +23,36 @@ class MockSMBO(SMBO):
 
 class TestPSMACFacade(unittest.TestCase):
     def setUp(self):
+        base_directory = os.path.split(__file__)[0]
+        base_directory = os.path.abspath(os.path.join(base_directory, "../../tests", ".."))
+        os.chdir(base_directory)
         self.output_dirs = []
         fn = "tests/test_files/spear_hydra_test_scenario.txt"
+        fn = "tests/test_files/test_deterministic_scenario.txt"
         self.scenario = Scenario(fn)
         self.scenario.limit_resources = True
 
     @patch("smac.facade.smac_ac_facade.SMBO", new=MockSMBO)
     def test_psmac(self):
-        # TODO: Fix tests
-        """
         import joblib
-        from smac.facade.experimental.psmac_facade import PSMAC
+        from smac.facade.psmac_facade import PSMAC
+        from smac.facade.smac_ac_facade import SMAC4AC
+        from smac.facade.smac_bb_facade import SMAC4BB
+        from smac.facade.smac_hpo_facade import SMAC4HPO
+        from smac.facade.smac_mf_facade import SMAC4MF
 
-        with joblib.parallel_backend("multiprocessing", n_jobs=1):
-            optimizer = PSMAC(self.scenario, n_optimizers=3, n_incs=2, validate=False)
-            incs = optimizer.optimize()
-            self.assertEqual(len(incs), 2)
-            optimizer = PSMAC(self.scenario, n_optimizers=1, n_incs=4, validate=False)
-            incs = optimizer.optimize()
-            self.assertEqual(len(incs), 2)
-            optimizer = PSMAC(self.scenario, n_optimizers=5, n_incs=4, validate=False)
-            incs = optimizer.optimize()
-            self.assertEqual(len(incs), 4)
-        """
-        ...
+        facades = [None, SMAC4AC, SMAC4BB, SMAC4HPO, SMAC4MF]
+        n_workers_list = [1, 2, 3, 4]
+        n_facades = len(facades)
+        target = {"x1": 7.290709845323256, "x2": 10.285684762665337}
+        for i, facade in enumerate(facades):
+            for j, n_workers in enumerate(n_workers_list):
+                idx = n_facades * i + j
+                with self.subTest(i=idx):
+                    with joblib.parallel_backend("multiprocessing", n_jobs=1):
+                        optimizer = PSMAC(self.scenario, facade_class=facade, n_workers=n_workers, validate=False)
+                        inc = optimizer.optimize()
+                        self.assertDictEqual(target, dict(inc))
 
     def tearDown(self):
         hydras = glob.glob1(".", "psmac*")