+++ changelog.md
@@ -4,13 +4,12 @@
 * [BOinG](https://arxiv.org/abs/2111.05834): A two-stage bayesian optimization approach to allow the 
 optimizer to focus on the most promising regions.
 * [TurBO](https://arxiv.org/abs/1910.01739): Reimplementaion of TurBO-1 algorithm.
-
+* Updated pSMAC: Can pass arbitrary SMAC facades now. Added example and fixed tests.
 
 ## Improvements
 * Enabled caching for multi-objectives (#872). Costs are now normalized in `get_cost` 
 or optionally in `average_cost`/`sum_cost`/`min_cost` to receive a single float value. Therefore,
 the cached cost values do not need to be updated everytime a new entry to the runhistory was added.
-
 
 ## Interface changes
 * We changed the location of gaussian processes and random forests. They are in the folders